requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
requests==2.31.0
openpyxl==3.1.2
mysqlclient==2.2.4